debezium/dbz#1417 Fix type of the average value

+++ debezium-connector-common/src/main/java/io/debezium/pipeline/meters/StreamingMeter.java
@@ -80,7 +80,7 @@ public class StreamingMeter implements StreamingMetricsMXBean, StreamingStatisti
     }
 
     @Override
-    public Long getMilliSecondsBehindSourceAverageValue() {
+    public Double getMilliSecondsBehindSourceAverageValue() {
         return lagBehindSourceMeasurement.getAverageValue();
     }
