+++ debezium-connector-common/src/main/java/io/debezium/pipeline/metrics/traits/StreamingStatisticsMXBean.java
@@ -15,7 +15,7 @@ public interface StreamingStatisticsMXBean {
 
     Long getMilliSecondsBehindSourceMaxValue();
 
-    Long getMilliSecondsBehindSourceAverageValue();
+    Double getMilliSecondsBehindSourceAverageValue();
 
     Double getMilliSecondsBehindSourceP50();
